Split test_export into better locations
